mineru/auth/requirements.txt
# 认证相关依赖
PyJWT==2.8.0
Flask==2.3.3
Werkzeug==2.3.7
